move lead component
