+++ kotlin-react-bootstrap/src/main/kotlin/react/bootstrap/content/typography/lead/Lead.kt
@@ -1,15 +1,12 @@
-package react.bootstrap.content.typography
+package react.bootstrap.content.typography.lead
 
 import kotlinx.html.classes
 import react.RBuilder
-import react.RHandler
 import react.RState
-import react.ReactElement
 import react.bootstrap.lib.bootstrap.ClassNames
 import react.bootstrap.lib.component.BootstrapComponent
 import react.bootstrap.lib.kotlinxhtml.loadGlobalAttributes
 import react.bootstrap.lib.react.rprops.WithGlobalAttributes
-import react.bootstrap.splitClassesToSet
 import react.dom.p
 
 class Lead : BootstrapComponent<Lead.Props, RState>() {
@@ -28,12 +25,3 @@ class Lead : BootstrapComponent<Lead.Props, RState>() {
 
     interface Props : WithGlobalAttributes
 }
-
-fun RBuilder.lead(classes: String? = null, block: RHandler<Lead.Props>): ReactElement =
-    child(Lead::class) {
-        attrs {
-            this.classes = classes.splitClassesToSet()
-        }
-
-        block()
-    }
