+++ site/src/main/kotlin/react/bootstrap/site/components/docs/content/typography/Lead.kt
@@ -1,7 +1,7 @@
 package react.bootstrap.site.components.docs.content.typography
 
 import react.RBuilder
-import react.bootstrap.content.typography.lead
+import react.bootstrap.content.typography.lead.lead
 import react.bootstrap.site.components.docs.fixings.SectionComponent
 import react.bootstrap.site.components.docs.fixings.codeExample
 import react.bootstrap.site.components.docs.fixings.liveExample
